% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate single-cell data}
\usage{
simulate(
  n_genes = 20000,
  n_cells = 100,
  rate = 0.1,
  shape = 0.1,
  dispersion = 0.1
)
}
\arguments{
\item{n_genes}{integer(1) the number of genes (rows) to simulate.}

\item{n_cells}{integer(1) the number of cells (columns) to simulate.}

\item{rate}{numeric(1) rate parameter of the `rgamma()` distribution.}

\item{shape}{numeric(1) shape parameter of the `rgamma()` distribution.}

\item{dispersion}{numeric(1) size (`1 / dispersion`) parameter of
the `rnbinom()` distribution.}
}
\value{
`simulate() returns a `n_genes x n_cells` matrix of
    simulated single-cell RNA-seq counts.
}
\description{
`simulate()` produces a genes x cells count matrix of
    simulated single-cell RNA-seq data. Gene expression is modelled
    using a gamma distribution. Counts are simulated using a
    negative binomial distribution.
}
\examples{
counts <- simulate()
dim(counts)
mean(counts == 0)  # fraction of '0' cells
range(counts)

}
