% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_factors.R
\name{size_factors}
\alias{size_factors}
\title{Calculate geometric mean-centered median scaled cell scaling
    factors.}
\usage{
size_factors(counts)
}
\arguments{
\item{counts}{matrix() of gene x cell RNA-seq counts.}
}
\value{
`size_factors()` returns a `numeric(ncol(counts))` vector
    of scaling factors.
}
\description{
`size_factors()` centers the log counts of each row
    (gene) by the row mean of the log counts. The finite centered
    values are then used to compute column-wise geometric median
    scaling factors.
}
\examples{
set.seed(123)
counts <- simulate()
size_factors <- size_factors(counts)
median(size_factors)      # approximately 1

}
